Update acquire file to convert numeric types  while keeping nan values.

+++ acquire.py
@@ -42,13 +42,13 @@ def get_riiid_data():
         df_filtered = df_merged.merge(df_questions, left_on='content_id', right_on='question_id', how='left')
         
         # Change the data types of numeric columns.
-        df_filtered.lecture_id = df_filtered.lecture_id.astype(np.int16)
-        df_filtered.tag = df_filtered.tag.astype(np.int8)
-        df_filtered.part_x = df_filtered.part_x.astype(np.int8)
-        df_filtered.part_y = df_filtered.part_y.astype(np.int8)
-        df_filtered.question_id = df_filtered.question_id.astype(np.int16)
-        df_filtered.bundle_id = df_filtered.bundle_id.astype(np.int16)
-        df_filtered.lecture_id = df_filtered.lecture_id.astype(np.int32)
+        df_filtered.lecture_id = df_filtered.lecture_id.astype('Int16')
+        df_filtered.tag = df_filtered.tag.astype('Int8')
+        df_filtered.part_x = df_filtered.part_x.astype('Int8')
+        df_filtered.part_y = df_filtered.part_y.astype('Int8')
+        df_filtered.question_id = df_filtered.question_id.astype('Int16')
+        df_filtered.bundle_id = df_filtered.bundle_id.astype('Int16')
+        df_filtered.lecture_id = df_filtered.lecture_id.astype('Int32')
 
         # Prefix part names with the originating dataframe name.
         df_filtered.rename(columns={'part_x': 'lecture_part',
